deterministic_kernel_flows/requirements.txt
importlib_metadata==8.7.0
ipykernel==7.1.0
matplotlib==3.10.7
pickleshare==0.7.5
pip==25.3
tensorflow_keras==0.1
tensorflow-probability==0.24.0
tf_keras==2.16.0
pytest==9.0.1
pandas==2.3.3